refactor to Properties/
