+++ src/components/FeaturePanel/Properties/Properties.tsx
@@ -1,12 +1,12 @@
-import { Typography } from '@material-ui/core';
 import React from 'react';
 import { FeaturedTags } from '../FeaturedTags';
 import { IdSchemeFields } from './IdSchemeFields';
 import { t } from '../../../services/intl';
 import { TagsTableInner } from './TagsTableInner';
 import { useFeatureContext } from '../../utils/FeatureContext';
+import { Subheading } from '../helpers/Subheading';
 
-export const TagsTable = ({ showTags }) => {
+export const Properties = ({ showTags }) => {
   const { feature } = useFeatureContext();
   const deleted = feature.error === 'deleted';
 
@@ -25,9 +25,7 @@ export const TagsTable = ({ showTags }) => {
       )}
       {showTags && (
         <>
-          <Typography variant="overline" display="block" color="textSecondary">
-            {t('featurepanel.all_tags_heading')}
-          </Typography>
+          <Subheading>{t('featurepanel.all_tags_heading')}</Subheading>
           <TagsTableInner
             tags={feature.tags}
             center={feature.center}
