+++ src/components/FeaturePanel/Properties/Table.tsx
