+++ src/components/FeaturePanel/Properties/TagsTableInner.tsx
