+++ src/components/FeaturePanel/helpers/Subheading.tsx
@@ -1,0 +1,8 @@
+import { Typography } from '@material-ui/core';
+import React from 'react';
+
+export const Subheading = ({ children }) => (
+  <Typography variant="overline" display="block" color="textSecondary">
+    {children}
+  </Typography>
+);
